Fixed TempZipFile saving when it shouldnt
Woops.

+++ src/main/java/net/glasslauncher/legacy/util/TempZipFile.java
@@ -79,37 +79,39 @@ public class TempZipFile {
     }
 
     public void close(boolean doSave) {
-        File original = new File(originalPath);
-        try {
-            original.delete();
-        } catch (Exception e) {
-            Main.getLogger().info("Failed to delete/check read permissions for \"" + originalPath + "\"");
-            e.printStackTrace();
-            return;
-        }
-        try {
-            List<String> args = new ArrayList<>();
-            args.add("jar");
-            args.add("cMf");
-            args.add(originalPath);
-            args.add("./*");
-            ProcessBuilder processBuilder = new ProcessBuilder(args);
-            processBuilder.redirectInput(ProcessBuilder.Redirect.INHERIT);
-            processBuilder.directory(new File(tempPath));
-            Process zipProcess = processBuilder.start();
-            zipProcess.waitFor(10, TimeUnit.MINUTES);
-            if (zipProcess.isAlive()) {
-                Main.getLogger().warning("Zip process has been running for longer than 10 minutes! Terminating process...");
-                Main.getLogger().warning("Target folder to zip was: \"" + tempPath + "\"");
-                zipProcess.destroy();
-                zipProcess.waitFor(10, TimeUnit.SECONDS);
-                if (zipProcess.isAlive()) {
-                    Main.getLogger().warning("Zip process has taken longer than 10 seconds to terminate! Force terminating process...");
-                    zipProcess.destroyForcibly();
-                }
+        if (doSave) {
+            File original = new File(originalPath);
+            try {
+                original.delete();
+            } catch (Exception e) {
+                Main.getLogger().info("Failed to delete/check read permissions for \"" + originalPath + "\"");
+                e.printStackTrace();
+                return;
             }
-        } catch (Exception e) {
-            e.printStackTrace();
+            try {
+                List<String> args = new ArrayList<>();
+                args.add("jar");
+                args.add("cMf");
+                args.add(originalPath);
+                args.add("./*");
+                ProcessBuilder processBuilder = new ProcessBuilder(args);
+                processBuilder.redirectInput(ProcessBuilder.Redirect.INHERIT);
+                processBuilder.directory(new File(tempPath));
+                Process zipProcess = processBuilder.start();
+                zipProcess.waitFor(10, TimeUnit.MINUTES);
+                if (zipProcess.isAlive()) {
+                    Main.getLogger().warning("Zip process has been running for longer than 10 minutes! Terminating process...");
+                    Main.getLogger().warning("Target folder to zip was: \"" + tempPath + "\"");
+                    zipProcess.destroy();
+                    zipProcess.waitFor(10, TimeUnit.SECONDS);
+                    if (zipProcess.isAlive()) {
+                        Main.getLogger().warning("Zip process has taken longer than 10 seconds to terminate! Force terminating process...");
+                        zipProcess.destroyForcibly();
+                    }
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
 
         try {
